requirements.txt
streamlit==1.28.2
openai==1.51.0
python-dotenv==1.0.0 